Crate luno

A Rust wrapper for the Luno API. 🦀🚀

In order to use some features of the Luno API you need to have an existing Luno account or you can sign up at https://luno.com/signup. Go to https://www.luno.com/wallet/security/api_keys to create a new API key with appropriate permissions.

Re-exports

pub use client::LunoClient;

pub use market::Currency;

pub use market::TradingPair;

pub use orders::LimitOrderType;

pub use orders::MarketOrderType;

Modules

accounts

client

lightning

market

orders

trades

transactions