Struct luno::client::LunoClient

pub struct LunoClient { /* fields omitted */ }

The top level client for interacting with the Luno API.

Methods

impl LunoClient

pub fn new(key: String, secret: String) -> LunoClient

pub fn get_ticker(
&self,
pair: TradingPair
) -> impl Future<Output = Result<Ticker, Error>> + '_

Returns the latest ticker indicators.

pub fn get_tickers(
&self
) -> impl Future<Output = Result<TickerList, Error>> + '_

Returns the latest ticker indicators from all active Luno exchanges.

pub fn get_orderbook_top(
&self,
pair: TradingPair
) -> impl Future<Output = Result<Orderbook, Error>> + '_

Returns a list of the top 100 bids and asks in the order book. Ask orders are sorted by price ascending. Bid orders are sorted by price descending. Orders of the same price are aggregated.

pub fn get_orderbook(
&self,
pair: TradingPair
) -> impl Future<Output = Result<Orderbook, Error>> + '_

Returns a list of all bids and asks in the order book. Ask orders are sorted by price ascending. Bid orders are sorted by price descending. Multiple orders at the same price are not aggregated.

Warning: This may return a large amount of data. Generally you should rather use get_orderbook_top or the Streaming API.

pub fn get_trades(
&self,
pair: TradingPair
) -> impl Future<Output = Result<TradeList, Error>> + '_

Returns a list of the most recent trades that happened in the last 24h. At most 100 results are returned per call.

pub async fn create_account<'_, '_>(
&'_ self,
currency: Currency,
name: &'_ str
) -> Result<Account, Error>

This request creates an Account for the specified currency. Please note that the balances for the Account will be displayed based on the asset value, which is the currency the Account is based on.

pub fn get_balances(
&self
) -> impl Future<Output = Result<BalanceList, Error>> + '_

The list of all Accounts and their respective balances for the requesting user.

pub fn get_transactions(
&self,
account_id: &str,
min_row: u64,
max_row: u64
) -> impl Future<Output = Result<TransactionList, Error>> + '_

Return a list of transaction entries from an account.

Transaction entry rows are numbered sequentially starting from 1, where 1 is the oldest entry. The range of rows to return are specified with the min_row (inclusive) and max_row (exclusive) parameters. At most 1000 rows can be requested per call. If min_row or max_row is non-positive, the range wraps around the most recent row. For example, to fetch the 100 most recent rows, use min_row=-100 and max_row=0.

pub fn get_pending_transactions(
&self,
account_id: &str
) -> impl Future<Output = Result<PendingTransactionList, Error>> + '_

Return a list of all transactions that have not completed for the Account.

Pending transactions are not numbered, and may be reordered, deleted or updated at any time.

pub fn list_orders(&self) -> ListOrdersBuilder

Get a list of the most recently placed orders. Note that list_orders() returns a ListOrdersBuilder that allows you chain pair and state filters onto your request.

pub fn limit_order(
&self,
pair: TradingPair,
r#type: LimitOrderType,
volume: f64,
price: f64
) -> PostLimitOrderBuilder

Create a new trade order.

Warning! Orders cannot be reversed once they have executed. Please ensure your program has been thoroughly tested before submitting orders.

If no base_account_id or counter_account_id are specified, your default base currency or counter currency account will be used. You can find your account IDs by calling get_balances().

pub fn market_order(
&self,
pair: TradingPair,
r#type: MarketOrderType,
volume: f64
) -> PostMarketOrderBuilder

Create a new market order.

A market order executes immediately, and either buys as much cryptocurrency that can be bought for a set amount of fiat currency, or sells a set amount of cryptocurrency for as much fiat as possible.

Warning! Orders cannot be reversed once they have executed. Please ensure your program has been thoroughly tested before submitting orders.

If no base_account_id or counter_account_id are specified, your default base currency or counter currency account will be used. You can find your account IDs by calling the get_balances().

pub async fn stop_order<'_, '_>(
&'_ self,
order_id: &'_ str
) -> Result<StopOrderResponse, Error>

Request to stop an order.

pub fn get_order(
&self,
order_id: &str
) -> impl Future<Output = Result<Order, Error>> + '_

Get an order by its ID.

pub fn list_own_trades(&self, pair: TradingPair) -> ListTradesBuilder

Returns a list of your recent trades for a given pair, sorted by oldest first. If before is specified, then the trades are returned sorted by most recent first.

type in the response indicates the type of order that you placed in order to participate in the trade. Possible types: BID, ASK.

If is_buy in the response is true, then the order which completed the trade (market taker) was a bid order.

Results of this query may lag behind the latest data.

pub fn get_fee_info(
&self,
pair: TradingPair
) -> impl Future<Output = Result<FeeInfo, Error>> + '_

Returns the fees and 30 day trading volume (as of midnight) for a given currency pair. For complete details, please see Fees & Features.

pub fn lightning_send(&self, payment_request: &str) -> LightningSendBuilder

Alpha warning! The Lightning API is still in Alpha stage. The risks are limited api availability and channel capacity.

Send Bitcoin over the lightning network from your Bitcoin wallet. Warning! Cryptocurrency transactions are irreversible.

Please ensure your program has been thoroughly tested before using this call.

pub fn lightning_receive(&self, amount: f64) -> LightningReceiveBuilder

Alpha warning! The Lightning API is still in Alpha stage. The risks are limited API availability and channel capacity.

Create a lightning invoice which can be used to receive BTC payments over the lightning network.

pub fn lookup_lightning_invoice(
&self,
id: i64
) -> impl Future<Output = Result<LightningInvoiceLookupResponse, Error>> + '_

Alpha warning! The Lightning API is still in Alpha stage. The risks are limited API availability and channel capacity.

Lookup the status of a lightning receive invoice.